Upgrade and fix upgrade process
Running the homeassistant command with the check_config script does not
seem to be enough to trigger an update of .HA_VERSION anymore. This
changed sometime after 0.60.0 and before the version that this commit
contains.  This is a hack to keep versions in sync and requires updating
before committing: fine for our current process but not as good if using
a CI server for deploys.